docker-compose small fix
Dockerfile

created the interface Contracts folder

created a Models folder. Sigh laravel 5.7 for not including this in their boilerplate

Migrated the UserRoles table. Implemented in a Contracts to Service flow and it is all getting called in UserRoleServicesProviders

working on testing UserRoleController

installing PSR_2 into my project

merge conflicts

Revert "created the interface Contracts folder"

This reverts commit af6612c983d646723e1d46a0eb18cd2ac705eaf4.

Revert "created a Models folder. Sigh laravel 5.7 for not including this in their boilerplate"

This reverts commit 0f6c642ef209f1fd765d63bd554c89e64bdadfab.

Revert "Migrated the UserRoles table. Implemented in a Contracts to Service flow and it is all getting called in UserRoleServicesProviders"

This reverts commit 7f23b23f5f8ffe8004f0b5854637f766716a5421.

Revert "working on testing UserRoleController"

This reverts commit 33a87d0efbc1747d18d7e17c5914c05e7ff491bc.

Revert "installing PSR_2 into my project"

This reverts commit dd08c3bd3837a10f8cbd36be795e16d29eb5a1c6.